PJFCB-22034 Set Kafka to 3.9.2, smallrye-reactive-messaging to 4.27.0, smallrye-common to 2.3.0
