+++ pom.xml
@@ -406,7 +406,7 @@
         <version.io.reactivex.rxjava3>3.1.8</version.io.reactivex.rxjava3>
         <version.io.smallrye.open-api>3.7.0</version.io.smallrye.open-api>
         <version.io.smallrye.reactive-utils>2.6.0</version.io.smallrye.reactive-utils>
-        <version.io.smallrye.smallrye-common>2.2.0</version.io.smallrye.smallrye-common>
+        <version.io.smallrye.smallrye-common>2.3.0</version.io.smallrye.smallrye-common>
         <version.io.smallrye.smallrye-config>3.4.4</version.io.smallrye.smallrye-config>
         <version.io.smallrye.smallrye-fault-tolerance>6.4.3</version.io.smallrye.smallrye-fault-tolerance>
         <version.io.smallrye.smallrye-health>4.0.4</version.io.smallrye.smallrye-health>
@@ -415,7 +415,7 @@
         <version.io.smallrye.smallrye-mutiny-vertx>3.8.0</version.io.smallrye.smallrye-mutiny-vertx>
         <version.io.smallrye.smallrye-mutiny-zero>1.0.0</version.io.smallrye.smallrye-mutiny-zero>
         <version.io.smallrye.smallrye-opentelemetry>2.6.0</version.io.smallrye.smallrye-opentelemetry>
-        <version.io.smallrye.smallrye-reactive-messaging>4.16.0</version.io.smallrye.smallrye-reactive-messaging>
+        <version.io.smallrye.smallrye-reactive-messaging>4.27.0</version.io.smallrye.smallrye-reactive-messaging>
         <version.io.undertow.jastow>2.2.7.Final</version.io.undertow.jastow>
         <version.io.vertx.vertx>4.5.30</version.io.vertx.vertx>
         <version.io.vertx.vertx-kafka-client>4.4.9</version.io.vertx.vertx-kafka-client>
@@ -453,7 +453,7 @@
         <version.org.apache.cxf.xjcplugins>4.1.3</version.org.apache.cxf.xjcplugins>
         <version.org.apache.httpcomponents.httpasyncclient>4.1.5</version.org.apache.httpcomponents.httpasyncclient>
         <version.org.apache.james.apache-mime4j>0.8.9</version.org.apache.james.apache-mime4j>
-        <version.org.apache.kafka>3.6.2</version.org.apache.kafka>
+        <version.org.apache.kafka>3.9.2</version.org.apache.kafka>
         <version.org.apache.lucene>9.8.0</version.org.apache.lucene>
         <version.org.apache.neethi>3.2.3</version.org.apache.neethi>
         <version.org.apache.qpid.proton>0.34.1</version.org.apache.qpid.proton>
